# gateway/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/apps/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Account API
    version: 0.0.1
paths:
    /account/v1/avatar:
        post:
            tags:
                - Account
            operationId: Account_UpdateAvatar
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateAvatarRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UpdateAvatarReply'
    /account/v1/login:
        post:
            tags:
                - Account
            operationId: Account_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LoginReply'
    /account/v1/register:
        post:
            tags:
                - Account
            operationId: Account_Register
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/RegisterRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/RegisterReply'
    /account/v1/search:
        get:
            tags:
                - Account
            operationId: Account_SearchUuid
            parameters:
                - name: uuid
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SearchUuidReply'
components:
    schemas:
        LoginReply:
            type: object
            properties:
                token:
                    type: string
                username:
                    type: string
                email:
                    type: string
                phone:
                    type: string
                avatar:
                    type: string
                nickname:
                    type: string
        LoginRequest:
            type: object
            properties:
                account:
                    type: string
                    description: user account, like username, email, phone number
                password:
                    type: string
                    description: md5(password+"piggytalk")
                reCaptchaToken:
                    type: string
        RegisterReply:
            type: object
            properties:
                token:
                    type: string
        RegisterRequest:
            type: object
            properties:
                username:
                    type: string
                password:
                    type: string
                    description: same as login request
                email:
                    type: string
                phone:
                    type: string
                avatar:
                    type: string
                nickname:
                    type: string
                reCaptchaToken:
                    type: string
        SearchUuidReply:
            type: object
            properties:
                uuid:
                    type: string
                avatar:
                    type: string
                nickname:
                    type: string
        UpdateAvatarReply:
            type: object
            properties:
                token:
                    type: string
                avatar:
                    type: string
        UpdateAvatarRequest:
            type: object
            properties:
                token:
                    type: string
                avatar:
                    type: string
tags:
    - name: Account